reorganize vuex state into modules
