+++ .prettierrc
@@ -1,5 +1,5 @@
 {
-  "printWidth": 100,
+  "printWidth": 80,
   "singleQuote": true,
   "useTabs": false,
   "semi": false
