+++ src/store/modules/index.js
@@ -1,0 +1,6 @@
+export { default as filter } from './filter'
+export { default as channel } from './channel'
+export { default as subject } from './subject'
+export { default as categorie } from './categorie'
+export { default as collaborator } from './collaborator'
+export { default as notification } from './notification'
